refactor: validate log_trace_height in config::verify and accept u32
Move the trace height validation into `config::verify` (the shared STARK
verification entry point) instead of the VM-specific verifier. Reject
proofs where `log_trace_height` exceeds the two-adic order of the
Goldilocks field (32). Also narrow the parameter type from `usize` to
`u32` to match `ExecutionProof::log_trace_height`.

+++ verifier/src/lib.rs
@@ -146,7 +146,7 @@ fn verify_stark(
             config::verify(
                 &config,
                 &air,
-                log_trace_height as usize,
+                log_trace_height,
                 &public_values,
                 var_len_public_inputs,
                 &proof_bytes,
@@ -157,7 +157,7 @@ fn verify_stark(
             config::verify(
                 &config,
                 &air,
-                log_trace_height as usize,
+                log_trace_height,
                 &public_values,
                 var_len_public_inputs,
                 &proof_bytes,
@@ -168,7 +168,7 @@ fn verify_stark(
             config::verify(
                 &config,
                 &air,
-                log_trace_height as usize,
+                log_trace_height,
                 &public_values,
                 var_len_public_inputs,
                 &proof_bytes,
@@ -179,7 +179,7 @@ fn verify_stark(
             config::verify(
                 &config,
                 &air,
-                log_trace_height as usize,
+                log_trace_height,
                 &public_values,
                 var_len_public_inputs,
                 &proof_bytes,
@@ -190,7 +190,7 @@ fn verify_stark(
             config::verify(
                 &config,
                 &air,
-                log_trace_height as usize,
+                log_trace_height,
                 &public_values,
                 var_len_public_inputs,
                 &proof_bytes,
